requirements.txt
openai
streamlit
langchain-openai
python-dotenv